backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.27
pydantic==2.5.3
pydantic[email]==2.5.3
python-dotenv==1.0.1
langchain==0.1.10
langchain-openai==0.0.6
langchain-community==0.0.18
openai==1.13.3
bcrypt==4.1.2
pytest==7.4.4
httpx==0.27.0
python-multipart==0.0.7